+++ NEWS.md
@@ -7,6 +7,7 @@ should be more robust (#397, thank you @logan-888!) (v1.4.8.02)
 - `ff_startpositions.sleeper_conn()` now properly handles leagues with flex
 positions that did not have dedicated posititions (#400, thank you @mcarman8!) 
 (v1.4.8.03)
+- `ff_scoring.mfl_conn()` now works better with new dplyr versions (#402) (v1.4.8.04)
 
 # ffscrapr 1.4.8